backend/requirements.txt
fastapi
uvicorn[standard]
neo4j
ollama
python-dotenv
pandas
